+++ java/src/main/java/io/cucumber/gherkin/GherkinDialect.java
@@ -39,7 +39,6 @@ public final class GherkinDialect {
     private final List<String> butKeywords;
     private final List<String> stepKeywords;
     private final Map<String, Set<StepKeywordType>> stepKeywordsTypes;
-    private final Map<String, StepKeywordType> stepKeywordsType;
 
     GherkinDialect(
             String language,
@@ -71,20 +70,9 @@ public final class GherkinDialect {
         this.thenKeywords = requireNonNull(thenKeywords);
         this.andKeywords = requireNonNull(andKeywords);
         this.butKeywords = requireNonNull(butKeywords);
-        
+
         this.stepKeywords = distinctKeywords(givenKeywords, whenKeywords, thenKeywords, andKeywords, butKeywords);
         this.stepKeywordsTypes = aggregateKeywordTypes(givenKeywords, whenKeywords, thenKeywords, andKeywords, butKeywords);
-        this.stepKeywordsType = defineSingleTypes();
-    }
-
-    private Map<String, StepKeywordType> defineSingleTypes() {
-        Map<String, StepKeywordType> stepKeywordsType = new HashMap<>();
-        for (Map.Entry<String, Set<StepKeywordType>> entry : stepKeywordsTypes.entrySet()) {
-            Set<StepKeywordType> stepKeywordTypes = entry.getValue();
-            StepKeywordType type = stepKeywordTypes.size() == 1 ? stepKeywordTypes.iterator().next() : StepKeywordType.UNKNOWN;
-            stepKeywordsType.put(entry.getKey(), type);
-        }
-        return stepKeywordsType;
     }
 
     @SafeVarargs
@@ -98,7 +86,7 @@ public final class GherkinDialect {
         sortedKeywords.sort(LONGEST_TO_SHORTEST_COMPARATOR);
         return unmodifiableList(sortedKeywords);
     }
-    
+
     private static Map<String, Set<StepKeywordType>> aggregateKeywordTypes(
             List<String> givenKeywords,
             List<String> whenKeywords,
@@ -157,10 +145,10 @@ public final class GherkinDialect {
 
     /**
      * Returns the {@link StepKeywordType StepKeywordTypes} for a given keyword
-     * 
-     * @deprecated use {{@link #getStepKeywordTypesSet(String)}} instead.
+     *
      * @param keyword to get the keyword type for
      * @return the keywords type
+     * @deprecated use {{@link #getStepKeywordTypesSet(String)}} instead.
      */
     @Deprecated
     public List<StepKeywordType> getStepKeywordTypes(String keyword) {
@@ -184,14 +172,6 @@ public final class GherkinDialect {
             throw new NoSuchElementException(String.format("'%s' is not part of this dialect", keyword));
         }
         return stepKeywordTypes;
-    }
-
-    StepKeywordType getStepKeywordType(String keyword) {
-        StepKeywordType stepKeywordType = stepKeywordsType.get(keyword);
-        if (stepKeywordType == null) {
-            throw new NoSuchElementException(String.format("'%s' is not part of this dialect", keyword));
-        }
-        return stepKeywordType;
     }
 
     public List<String> getBackgroundKeywords() {