+++ java/src/main/java/io/cucumber/gherkin/GherkinTokenMatcher.java
@@ -1,18 +1,21 @@
 package io.cucumber.gherkin;
 
+import io.cucumber.gherkin.Parser.TokenMatcher;
+import io.cucumber.messages.types.Location;
+import io.cucumber.messages.types.StepKeywordType;
+
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
+import java.util.function.Function;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
-
-import io.cucumber.gherkin.Parser.TokenMatcher;
-import io.cucumber.messages.types.StepKeywordType;
 
 import static io.cucumber.gherkin.GherkinLanguageConstants.COMMENT_PREFIX_CHAR;
 import static io.cucumber.gherkin.GherkinLanguageConstants.DOCSTRING_ALTERNATIVE_SEPARATOR;
 import static io.cucumber.gherkin.GherkinLanguageConstants.DOCSTRING_SEPARATOR;
 import static io.cucumber.gherkin.GherkinLanguageConstants.TABLE_CELL_SEPARATOR;
 import static io.cucumber.gherkin.GherkinLanguageConstants.TAG_PREFIX_CHAR;
-import static io.cucumber.gherkin.GherkinLanguageConstants.TITLE_KEYWORD_SEPARATOR_LENGTH;
 import static io.cucumber.gherkin.Locations.COLUMN_OFFSET;
 import static io.cucumber.gherkin.Locations.atColumn;
 import static io.cucumber.gherkin.Parser.TokenType;
@@ -20,22 +23,22 @@ import static io.cucumber.gherkin.Parser.TokenType;
 final class GherkinTokenMatcher implements TokenMatcher {
 
     private static final Pattern LANGUAGE_PATTERN = Pattern.compile("^#\\s*language\\s*:\\s*([a-zA-Z\\-_]+)\\s*$");
-    private final GherkinDialectProvider dialectProvider;
-    private GherkinDialect currentDialect;
+    private static final String DEFAULT_LANGUAGE = "en";
+    private final String defaultLanguage;
+    // Expect at most two languages, the default language and one other
+    private final Map<String, KeywordMatcher> activeKeywordMatchers = new HashMap<>(2);
+    private String currentLanguage;
+    private KeywordMatcher currentKeywordMatcher;
     private String activeDocStringSeparator = null;
     private int indentToRemove = 0;
 
-    GherkinTokenMatcher(GherkinDialectProvider dialectProvider) {
-        this.dialectProvider = dialectProvider;
+    GherkinTokenMatcher(String defaultLanguage) {
+        this.defaultLanguage = defaultLanguage;
         reset();
     }
 
     GherkinTokenMatcher() {
-        this(new GherkinDialectProvider());
-    }
-
-    GherkinTokenMatcher(String defaultDialectName) {
-        this(new GherkinDialectProvider(defaultDialectName));
+        this(DEFAULT_LANGUAGE);
     }
 
     @Override
@@ -43,7 +46,19 @@ final class GherkinTokenMatcher implements TokenMatcher {
         // TODO performance: reset() is called once in the constructor and once for each file (Parser.parse()). It could be called only once, but there is no measurable impact with the profiler
         activeDocStringSeparator = null;
         indentToRemove = 0;
-        currentDialect = dialectProvider.getDefaultDialect();
+        setLanguageMatched(defaultLanguage, null);
+    }
+
+    private void setLanguageMatched(String language, Location location) {
+        if (language.equals(currentLanguage)) {
+            return;
+        }
+        KeywordMatcher keywordMatcher = activeKeywordMatchers.computeIfAbsent(language, KeywordMatchers::of);
+        if (keywordMatcher == null) {
+            throw new ParserException.NoSuchLanguageException(language, location);
+        }
+        currentLanguage = language;
+        currentKeywordMatcher = keywordMatcher;
     }
 
     private void setTokenMatched(Token token, TokenType matchedType, String text, String keyword, int indent, StepKeywordType keywordType, List<GherkinLineSpan> items) {
@@ -52,7 +67,7 @@ final class GherkinTokenMatcher implements TokenMatcher {
         token.keywordType = keywordType;
         token.matchedText = text;
         token.matchedItems = items;
-        token.matchedGherkinDialect = currentDialect;
+        token.matchedLanguage = currentLanguage;
         token.matchedIndent = indent;
         token.location = atColumn(token.location, token.matchedIndent + COLUMN_OFFSET);
     }
@@ -98,9 +113,7 @@ final class GherkinTokenMatcher implements TokenMatcher {
         if (matcher.matches()) {
             String language = matcher.group(1);
             setTokenMatched(token, TokenType.Language, language, null, token.line.getIndent(), null, null);
-
-            currentDialect = dialectProvider.getDialect(language)
-                    .orElseThrow(() -> new ParserException.NoSuchLanguageException(language, token.location));
+            setLanguageMatched(language, token.location);
             return true;
         }
         return false;
@@ -118,41 +131,39 @@ final class GherkinTokenMatcher implements TokenMatcher {
 
     @Override
     public boolean match_FeatureLine(Token token) {
-        return matchTitleLine(token, TokenType.FeatureLine, currentDialect.getFeatureKeywords());
+        return matchTitleLine(token, TokenType.FeatureLine, currentKeywordMatcher::matchFeatureKeyword);
     }
 
     @Override
     public boolean match_RuleLine(Token token) {
-        return matchTitleLine(token, TokenType.RuleLine, currentDialect.getRuleKeywords());
+        return matchTitleLine(token, TokenType.RuleLine, currentKeywordMatcher::matchRuleKeyword);
     }
 
     @Override
     public boolean match_BackgroundLine(Token token) {
-        return matchTitleLine(token, TokenType.BackgroundLine, currentDialect.getBackgroundKeywords());
+        return matchTitleLine(token, TokenType.BackgroundLine, currentKeywordMatcher::matchBackgroundKeyword);
     }
 
     @Override
     public boolean match_ScenarioLine(Token token) {
-        return matchTitleLine(token, TokenType.ScenarioLine, currentDialect.getScenarioKeywords()) ||
-                matchTitleLine(token, TokenType.ScenarioLine, currentDialect.getScenarioOutlineKeywords());
+        return matchTitleLine(token, TokenType.ScenarioLine, currentKeywordMatcher::matchScenarioKeyword);
     }
 
     @Override
     public boolean match_ExamplesLine(Token token) {
-        return matchTitleLine(token, TokenType.ExamplesLine, currentDialect.getExamplesKeywords());
+        return matchTitleLine(token, TokenType.ExamplesLine, currentKeywordMatcher::matchExampleKeyword);
     }
 
-    @SuppressWarnings("ForLoopReplaceableByForEach") // classic 'for' loop is ~2x faster than 'for-each'
-    private boolean matchTitleLine(Token token, TokenType tokenType, List<String> keywords) {
-        for (int i = 0, keywordsSize = keywords.size(); i < keywordsSize; i++) {
-            String keyword = keywords.get(i);
-            if (token.line.startsWithTitleKeyword(keyword)) {
-                String title = token.line.substringTrimmed(keyword.length() + TITLE_KEYWORD_SEPARATOR_LENGTH);
-                setTokenMatched(token, tokenType, title, keyword, token.line.getIndent(), null, null);
-                return true;
-            }
+    private boolean matchTitleLine(Token token, TokenType tokenType, Function<GherkinLine, KeywordMatcher.Match> matcher) {
+        KeywordMatcher.Match match = matcher.apply(token.line);
+        if (match == null) {
+            return false;
         }
-        return false;
+        String keyword = match.getKeyword();
+        int keywordLength = match.getKeywordLength();
+        String title = token.line.substringTrimmed(keywordLength);
+        setTokenMatched(token, tokenType, title, keyword, token.line.getIndent(), null, null);
+        return true;
     }
 
     @Override
@@ -185,16 +196,16 @@ final class GherkinTokenMatcher implements TokenMatcher {
 
     @Override
     public boolean match_StepLine(Token token) {
-        List<String> keywords = currentDialect.getStepKeywords();
-        for (String keyword : keywords) {
-            if (token.line.startsWith(keyword)) {
-                String stepText = token.line.substringTrimmed(keyword.length());
-                StepKeywordType keywordType = currentDialect.getStepKeywordType(keyword);
-                setTokenMatched(token, TokenType.StepLine, stepText, keyword, token.line.getIndent(), keywordType, null);
-                return true;
-            }
+        KeywordMatcher.StepMatch match = currentKeywordMatcher.matchStepKeyword(token.line);
+        if (match == null) {
+            return false;
         }
-        return false;
+        String keyword = match.getKeyword();
+        int keywordLength = match.getKeywordLength();
+        StepKeywordType keywordType = match.getKeywordType();
+        String stepText = token.line.substringTrimmed(keywordLength);
+        setTokenMatched(token, TokenType.StepLine, stepText, keyword, token.line.getIndent(), keywordType, null);
+        return true;
     }
 
     @Override